消息通知问题

+++ divine-modules/divine-system/src/main/java/com/divine/system/domain/entity/SysNoticeRule.java
@@ -34,10 +34,10 @@ public class SysNoticeRule implements Serializable {
     /**
      * 1用户 2岗位
      */
-    private Integer targetType;
+    private Long targetType;
     /**
      * 用户ID/岗位ID
      */
-    private Integer targetId;
+    private Long targetId;
 
 }
